news_collection/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_83242e7e-f82d-4a71-8ef2-9d71d212d249gu_wxmeq/croots/recipe/certifi_1655968827803/work/certifi
Django==3.2
djangorestframework==3.13.1
pytz==2022.2.1
sqlparse==0.4.2
django-elasticsearch-dsl==7.2.1
celery==5.2.7
django-celery-beat==2.3.0